Adjusts in the documentation
